WIP - factor out Message
